LayoutTests/imported/w3c/web-platform-tests/fetch/api/basic/request-forbidden-headers.any-expected.txt

CONSOLE MESSAGE: line 4: ReferenceError: Can't find variable: RESOURCES_DIR

Harness Error (FAIL), message = ReferenceError: Can't find variable: RESOURCES_DIR
